+++ CHANGELOG.md
@@ -88,6 +88,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Replace `SPECIE_DOC` with title's `kind`
 - Rewrite ShippingFile Cnab* views
 - Update [aryelgois/Medools] config file
+- BankBillet and ShippingFile Controllers and Views use FilePack
 
 ### Deprecated